requirements.txt
Flask==2.3.2
flask_babel==4.0.0
Flask_JWT_Extended==4.7.1
Flask_Migrate==4.1.0
Flask_SQLAlchemy==2.5.1
python-dotenv==1.1.0
Requests==2.32.3
SQLAlchemy==1.4.37
Werkzeug==2.3.3
